#171 Support JDK 16
